requirements.txt
django==5.0.1
djangorestframework==3.14.0
zappa==0.58.0
python-dotenv==1.0.0
pynamodb==6.0.0